# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "erdsea-api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "paths": {
        "/accounts/{userAddress}": {
            "get": {
                "description": "Retrieves an account by an elrond user address (erd1...)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account by user address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user address",
                        "name": "userAddress",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Account"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Sets an account settable information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Set account information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user address",
                        "name": "userAddress",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "account info",
                        "name": "setAccountRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.SetAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Account"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/collections/{collectionName}": {
            "get": {
                "description": "Retrieves a collection by its name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Gets collection by name.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection name",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Collection"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Sets info for a collection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collections"
                ],
                "summary": "Set collection info.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection name",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "collection info",
                        "name": "updateCollectionRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.Collection"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/search/accounts/{accountName}": {
            "get": {
                "description": "Searches for accounts by name. Cached for 20 minutes. Limit 5 elements.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search accounts by name.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search string",
                        "name": "accountName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Account"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/search/collections/{collectionName}": {
            "get": {
                "description": "Searches for collections by name. Cached for 20 minutes. Limit 5 elements.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search collections by name.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search string",
                        "name": "collectionName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Collection"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/search/{searchString}": {
            "get": {
                "description": "Searches for collections by name and accounts by name. Cached for 20 minutes. Limit 5 elements for each.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "General search by string.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search string",
                        "name": "searchString",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GeneralSearchResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "data.Account": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instagramLink": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "twitterLink": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "data.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "data.Collection": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "integer"
                },
                "creatorID": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "discordLink": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instagramLink": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "telegramLink": {
                    "type": "string"
                },
                "tokenID": {
                    "type": "string"
                },
                "twitterLink": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "handlers.GeneralSearchResponse": {
            "type": "object"
        },
...
        "services.SetAccountRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "instagramLink": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "twitterLink": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "services.UpdateCollectionRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "discordLink": {
                    "type": "string"
                },
                "instagramLink": {
                    "type": "string"
                },
                "telegramLink": {
                    "type": "string"
                },
                "twitterLink": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        }
    }
}